Update Stream Analytics SDK
-Copied over new generated files that updates the SDK to the latest
version which added the diagnostic property to inputs and outputs.

If there is a problem with an input or output, the user can make a
ListInputInJob or ListOutputInJob method call with the
PropertiesToSelect parameter set to "*" (asterisk). This will return all
the inputs or outputs in the job with the diagnostic property set if
there is a problem with that particular input or output.

-Updated SDK NuGet version number to 1.4.1
-Updated AssemblyFileVersion to 1.4.1.0