Search SDK: Removing dependency on JsonConvert.DefaultSettings
This fixes issue: https://github.com/Azure/azure-sdk-for-net/issues/1438
In one case, we were calling JsonConvert.DeserializeObject without passing in
JsonSerializerSettings. However, just adding the settings wasn't sufficient to
completely fix the issue because JsonConvert merges some of the settings with
the defaults. The solution was to roll our own DeserializeObject and
SerializeObject methods that ignore JsonConvert.DefaultSettings.