Search SDK: Customizing important model classes
This change moves the boundary between generated and custom code in the Search
SDK. A few classes that used to be generated will be maintained by hand from
now on, at least in the .NET SDK.

Some background: The Search API supports two query operations in the data
plane: Search and Suggest, each of which can be invoked with either GET or
POST, and each of which has generic and non-generic C# methods (for customers
that use strongly-typed or dynamic models in their applications,
respectively). Currently the code for these operations is all custom.
Eventually, we will try to generate code for the non-generic part of the data
plane and implement the generic part in terms of it, but that's still a few
months off. No matter what implementation strategy we use, the
SearchParameters and SuggestParameters classes need to remain stable for
usability reasons (see ScoringParameter as an example of why this is useful).

The other classes changed in this pack are workarounds for the current lack of
client property renaming in AutoRest. See this issue for details:

https://github.com/Azure/autorest/issues/316