Search SDK: Normalizing usage of nullable types
The goal of this change is to ensure that all model types use nullable types
consistently. Currently AutoRest uses nullable types for all C# primitive
types, but this is unncessary for required properties. See tracking issue:

https://github.com/Azure/autorest/issues/375

For this reason, we have to exclude a few more model types until the above
issue is fixed.