Search SDK: Updating to VS2015 and making code portable to .NET Core
Due to a possible bug in the DNX toolchain, I can't actually get the SDK to
build for DNX, PCL, and .NET 4.5 at the same time. Rather than block on this,
I've backed out the project file changes but kept most of the changes to the
source code. Once the bug is fixed, I'll re-introduce the new project files.