Changeset includes:
1] Introduction of the InMage provider
2] Renaming the VmWareAzureV2 provider to InMageAzureV2 provider.
3] Updating the tests to use the new contracts.

Please Note: Since the service side changes for the new contracts are not checked in, the service is still using the old contracts. Therefore, I have retained the old contracts in this branch as well. Please modify the portal to use the new contracts only after the service has been updated.