Search SDK: Simplifying approach to supporting non-nullable primitive types
This is part of an overall effort to integrate the custom serialization in the
Azure Search SDK with the boilerplate serialization settings generated by
AutoRest. Removing ValueTypePreservingContractResolver will make this a lot
simpler, while using NullValueHandling instead of DefaultValueHandling ensures
that the Search SDK will continue to support non-nullable primitive types for
document fields.