Search SDK: Add support for serializing SearchContinuationToken
Many customers use Azure Search from a middle tier, but need to pass search
results to a browser-based client. A frequent request is to be able to
serialize DocumentSearchResult directly, including its ContinuationToken
property. This allows the middle tier to remain stateless between search
requests, since the continuation token is externalized and passed between
the browser and the middle tier.