Search SDK: Ensuring that all URL paths are in OData format
Azure Search allows normal path syntax for keys, but sometimes this can lead
to ambiguity between operation names and resource names, like index names.
For example, naming an index "stats" would cause problems because "stats" is
a reserved action name in the REST API. We're making our routing smarter in
the API to make such clashes less likely, but we're updating the SDK just to
be on the safe side.