Search SDK: Allowing redundant CreateOrUpdate methods to be generated
This is related to the following issue: https://github.com/Azure/autorest/issues/382

Azure Search doesn't support renaming resources, so strictly speaking the
overloads that take an extra name parameter are not required, but it doesn't
hurt anything to have them there in case we support rename in the future. At
least by moving our preferred overloads of CreateOrUpdate to custom code, we
can stop fighting the code generator.