Search SDK: Scoring parameter fix, part 1
Scoring parameter syntax has changed in the Azure Search REST API. Instead of
name:value pairs, they are now name-values pairs, where values are comma
separated and can optionally be quoted if they contain commas.

To make this invisible to users of the SDK, we're adding a new constructor
and Values property to the ScoringParameter class and marking the old
constructor and Value property as Obsolete.

At least one more commit will be needed to update session records. This is
currently blocked until we can obtain credentials for our test environment.