Search SDK: Introducing internal IDocumentsProxyOperations
AutoRest does not support the advanced features we need for a good data plane
client, like C# generics. Up until now we've worked around this by
hand-writing most of the methods of IDocumentsOperations. However, this must
change over time if we're to easily port our SDK to other languages.

The approach we will take starting with this change is to wrap AutoRest-
generated methods wherever possible. The generated methods will go in a new
interface, IDocumentsProxyOperations, which we force to be internal via the
Fix-GeneratedCode PowerShell script. Our public methods will be hand-written
wrappers that still live in IDocumentsOperations.

In this change we just establish the basic pattern with the Count method,
which previously had the wrong return type (long? instead of long) due to
AutoRest's lack of support for nullable type semantics.