Adding new CreateOrUpdateWithTimeConversion method
Adding a new method to ScheduleOperations to call the new overloaded PUT
web service URL that will convert the StartTime and ExpiryTime
properties to time specified in the TimeZone property. The Offset value
of those properties will be ignored and a new offset will be specified
for the times that corresponds to the time zone specified.