Search SDK: Factoring out base class for AnalyzerName and DataType
Also changing the way that such ExtensibleEnums are serialized. Now it is
possible to use [JsonConverter(typeof(ExtensibleEnumConverter))] on the model
type itself, which will allow us to use such types in generated code. This is
important for custom analyzers, which introduces several new extensible enum
types: TokenizerName, TokenFilterName, CharFilterName, and RegexFlags.