Search SDK: Fixing build issue caused by .NET Core RTM upgrade (#2150)
The build issue in Search.Tests went uncaught for a long time because it was
not being included in the command-line build. This was due to the behavior of
the custom FilterOutAutoRestLibraries build task, which expects a 1-1
correspondence of AutoRest libraries to test projects. To fix this, we have
(finally) split our very-unfinished preview Management library into its own
project, Microsoft.Azure.Management.Search.