Search SDK: Adding ability to change indexes on a SearchIndexClient
The new TargetDifferentIndex() method allows you to change the index to which
a SearchIndexClient will send its requests. This is useful for scenarios where
you have many indexes and need to pool SearchIndexClients to avoid running out
of TCP connections. This method is NOT thread-safe and should be used with due
care.

This change is a fix for the following reported issue: https://github.com/Azure/azure-sdk-for-net/issues/2225