Search SDK: Replacing comparison test code with ModelComparer
ModelComparer is a new class that will make it easier to add new tests in the
future (especially for custom analyzers, which introduces dozens of new model
classes). It uses reflection to recursively compare complex object structures.
The rules it uses are a sort of compromise between .NET type semantics and the
semantics we expect from an OData JSON-based REST API.

This change also enables XML documentation for the Microsoft.Azure.Search
project, and re-arranges the folder structure for some model classes.