Search SDK: Fixing datetime deserialization issue
Fix for issue https://github.com/Azure/azure-sdk-for-net/issues/2434

Currently, document payloads containing arrays of strings where some of the
strings are formatted as ISO datetime values cause exceptions to be thrown on
deserialization. This is because JSON.NET automatically deserializes such
values to DateTimeOffset in the absence of type information, and then we
subsequently try to convert the DateTimeOffset value to System.String, which
fails.

The fix for now is to be more tolerant of values that JSON.NET does not parse
as strings. This is not a great solution since it doesn't guarantee full
fidelity of all types (for example, if you try to round-trip a date time
value, it might not work due to different precision in the fractional seconds
part). However, a complete solution to this problem requires type information,
which we'd need to add to the API responses. This is on our long-term radar,
but it will be a while before we solve it.

I've also added new test coverage for DocumentConverter to pin the behavior
around deserialization of responses in the absence of type information.