Search SDK: Fix for equality comparison issues with DataType
Added == and != operators for all ExtensibleEnums. One side effect of this is
that it became possible to compare enums of different types, which doesn't
seem safe. The solution is to make the enum-to-string conversion explicit,
while defining an implicit string-to-enum conversion for usability (especially
in custom analyzers scenarios).

This is a breaking change, so it will target the next RC build of the SDK.