Search SDK: Merging 3.0-rc changes into AutoRest branch (#2590)
* Search SDK: Fix for equality comparison issues with DataType

Added == and != operators for all ExtensibleEnums. One side effect of this is
that it became possible to compare enums of different types, which doesn't
seem safe. The solution is to make the enum-to-string conversion explicit,
while defining an implicit string-to-enum conversion for usability (especially
in custom analyzers scenarios).

This is a breaking change, so it will target the next RC build of the SDK.

* Search SDK: Added some missing implicit conversion operators

* Search SDK: Adding implicit conversion to BlobExtractionMode, plus cleanup

* Search SDK: Updating API version and all documentation links

* Search SDK: Bumping major version number to 3.0-rc

* Search SDK: Re-record all tests for new API version

Also fixed a bunch of test failures. The causes were:

- Some analysis components were using old versions that are disallowed in the
new API version.
- Some analysis components were missing.
- The continuation serialization tests still had the old API version
hard-coded.
- One of the Analyze API tests needed to omit a few analysis components that
currently trigger known validation bugs. This test should be changed back
once the bugs are fixed.

* Search SDK: Removing features that are deprecated or not GA yet

The following were deprecated and have been removed:

- The single-valued constructor of ScoringParameter
- The Value property of ScoringParameter
- The RequestId property of SearchRequestOptions

Other members marked as [Obsolete] were not removed because they are required
for backwards compatibility with deserialized response payloads from the API.

Support for JSON and CSV parsing has also been removed from
IndexingParametersExtensions since these features of the Blob Indexer are not
GA yet.

* Search SDK: Adding back some obsolete methods

They were removed too soon.

* Search SDK: Removing unused session records

* Search SDK: Updating a few URLs from azure.com to docs.microsoft.com