Search SDK: Fixing unit tests that were broken by VS2017 migration
As part of moving to VS2017, the Search SDK is using a newer version of the
ClientRuntime. Two of the unit tests were pinning constructor validation
behavior of ServiceClient, which seems to have changed.

Also, the move to VS2017 brought back the default behavior of xunit running
all tests in parallel. This won't work, since the HttpMockServer is
apparently not thread-safe. I disabled parallel test execution to avoid this
problem. All tests now pass in playback mode when run from VS2017.