Search SDK: Fixing unit tests that were broken by VS2017 migration
As part of moving to VS2017, the Search SDK is using a newer version of the
ClientRuntime. Two of the unit tests were pinning constructor validation
behavior of ServiceClient, which seems to have changed.

Also, the move to VS2017 required using a newer version of the Resource
Manager SDK, which required re-recording all mocks.

Finally, for the data plane SDK, we upgraded to Microsoft.Spatial 7.2.0, which
is compatible with VS2017 and .NET Core.