Update example to use AzureCredentials
According to https://github.com/Azure/azure-sdk-for-net/blob/4c6f0794b9a27a386fc43ddd5396dcf327b9f340/AUTH.md

Using a credential file is not necessary the best practice for authenticating against Azure.  Programmatic keys are a good choice for any "API first" documentation, so this update makes that change.  

It also includes a comment with the entire namespace of the `AzureCredentialsFactory` type because it exists in multiple namespaces.  They may all be valid, but this one works and a comment may lower the barrier of entry if someone were to hit that problem.