Search SDK: Re-applying changes from search-4x-preview branch to vs17Dev
I tried rebasing the search-4x-preview branch to vs17Dev, but the results were
very messy, so instead this is just a squash-merge of all the changes. This
should differ from search-4x-preview only in the project files and some minor
test changes to account for behavioral changes in the latest version of
ClientRuntime.