Add Azure Stream Analytics SDK project and code
- Added StreamAnalytics folder
- Added csproj and sln files to set up project infrastructure
- Added code generated from Swagger specs found here:
https://github.com/Azure/azure-rest-api-specs/tree/master/arm-streamanalytics
- Bumped Assembly version to 2.0.0.0 to signify breaking changes are
accompanied with this new SDK
- Added test infrastructure and tests cases to test the generated code
for major user scenarios
- Ran tests and copied over the session records for playback

NOTE: Code for the Azure Stream Analytics SDK was previously generated
using Hydra/Hyak. Now that the code is being generated using
AutoRest/Swagger specs, the NuGet Package published that will contain
this code will introduce breaking changes compared to the previous NuGet
Package published. As a result, the version of the SDK was bumped up to
2.0.0.0 from 1.8.3.0.