Closing amqp-session when creating link fails. (#325)
We create a new amqp-session on which we open links. But if link creation fails, we should close the session as it is not re-used again. These sessions are mapped to a session handle table on backend which keeps track of these open sessions. Once we hit 5000 open sessions, backend throws QuotaExceededException.
Fixes #237