[Azure Search] Fixing some long-lived bugs in the data plane SDK (#4036)
* [Azure Search] Restoring build script and bumping data plane SDK version

The version bump is in preparation for a series of bug fixes that have been
pending for a long time. The build script change was taken from the old
abandoned vs17Dev branch.

* [Azure Search] Fix for data plane SDK issue: 3309

The FieldBuilder class in the Search .NET SDK now ignores properties marked
with JsonIgnore.

* [Azure Search] GetClient propagates SearchServiceClient's HttpClientHandler

Fixes an issue whereby setting the HttpClientHandler when creating a
SearchServiceClient has no effect on the SearchIndexClient returned by
Indexes.GetClient. This would break scenarios such as enabling GZip
compression, for example.

Now the HttpClientHandler propagates, but the DelegatingHandlers don't since
they can't be cloned and are mutated by the ServiceClient base class
constructor.