[ClientRuntime] Fixing spurious test failures in JsonSerializationTests (#4060)
Two of the JsonSerializationTests modify JsonConvert.DefaultSettings. They do
this in order to validate that SafeJsonConvert does not rely on these
settings.

Before this change, these tests would install a custom JSON converter and
contract resolver that always throw. Unfortunately, other aspects of the test
infrastructure rely on DefaultSettings, so this would sometimes result in
unexpected exceptions being thrown.

The solution is to weaken the effect of changing DefaultSettings. Now, the
SafeJsonConvert tests install a custom converter for their own Model class
that produces incorrect results. Because it's targeted at that class
specifically, it should have no side effects on other code that relies on
DefaultSettings.