[Azure Search] Merge 4.x preview code to psSdkJson6 branch and bump version (#4062)
* Search SDK: Re-applying changes from search-4x-preview branch to vs17Dev

I tried rebasing the search-4x-preview branch to vs17Dev, but the results were
very messy, so instead this is just a squash-merge of all the changes. This
should differ from search-4x-preview only in the project files and some minor
test changes to account for behavioral changes in the latest version of
ClientRuntime.

* Search SDK: Fixing failing tests

* Search SDK: Re-recording mocks for the preview SDKs

* Search SDK: Updating copyright dates

* Search SDK: Rename Management SDK project and update build script (#3236)

* Renaming Management.Search to Microsoft.Azure.Management.Search

The package and assembly names were accidentally changed during the VS2017
migration. This fixes it.

* Search SDK: Updating Build-SearchPackages script

* [Azure Search] Bumping preview SDK version to 4.0.2-preview

This new version includes some minor bug fixes that just shipped in the 3.0.5
GA SDK.