[Azure Search] Splitting the data plane into four packages
What's motivating this change is a recognition that the data plane SDK is
really fulfilling three scenarios:

 1. Actually implementing search features in applications (full-text search,
    paging, faceted navigation, etc.)
 2. Indexing data via the push API
 3. Automating the management of indexes, synonyms, indexers, data sources,
    and other service-level resources.

Scenarios 1 and 2 are what you'd normally think of as "data plane". Scenario
3 is closer to "control plane" or management plane, or even "metadata plane"
since managing index schemas in particular is still pretty close to the data.

The split goes something like this:
 - Microsoft.Azure.Search.Data: Fulfills scenarios 1 and 2 above. This covers
   everything under the /docs resource. Currently most of this is custom code.
 - Microsoft.Azure.Search.Service: Fulfills scenario 3 above. This covers all
   top-level resources of the Azure Search Service REST API. This is mostly
   generated code, and our eventual goal is for nearly all of it to be
   generated.
 - Microsoft.Azure.Search.Common: Utilities common to Data and Service.
 - Microsoft.Azure.Search: A stub package that depends on all the others,
   plus defines the Indexes.GetClient() method that acts as a bridge from
   SearchServiceClient to SearchIndexClient. This package provides a
   convenient upgrade path for existing SDK users.

Eventually we're thinking of using Microsoft.Azure.Search.Service in
conjunction with Microsoft.Azure.Management.Search to implement a rich set
of PowerShell CmdLets that can manage indexes, data sources, indexers, etc.
in addition to the ARM-level resources.