[Azure Search] Updating preview SDK version to 6.0-preview
This is necessary because the new Autocomplete API requires adding a new
method to an existing interface, which is technically a breaking change in
.NET. Also, version 5 now points to a different REST API version, and we don't
want to allow mixing versions in the same application.

This also updates the release notes to reflect that Synonyms is out of preview
and Autocomplete is in.