Add support for binary serialization/deserialization for HttpRecorder
The HttpRecorder did not correctly play back binary content.  The propsed fix is to introduce
new fields in `RecordEntry` named `IsRequestBodyBinary`/`IsResponseBodyBinary`, and use base64
encoding/decoding when the payload is deemed to be binary.  Currently the test for binary-ness
is a simple MIME type regex match.

Note that this PR includes tests to confirm the new behavior.  It will failed to build without
hacking around the fact that the test code depends on the NuGet-installed Microsoft.Azure.Test.HttpRecorder.dll.
If it is preferable to split the PR so you can update NuGet first without the tests failing, please
let me know.