Gracefully handle RenewMessageLockTask (#548)
In the MessageReceivePump class the RenewMessageLockTask is now canceled
gracefully instead of raise a TaskCancelledException and the catch it
again.

The code now doesn't await the canceled task, but await a new task which
just returns the canceled task. And check if the delay task is canceled
to determine if the RenewMessageLockTask should be completed or not.

Fixes #547.