Fix equality check for QueueDescription (#629)
The equality check between a manually created created instance of the QueueDescription class and an instance parsed from the XML response by the ManagementClient will not be considered equal. This is an issue when doing operations like create a queue if it is non existing.
The issue is originated in the property AuthorizationRules that was setting the underlying field if it was null. When an instance is being parsed from an non-existing set of rules in the XML response this property is set to an empty AuthorizationRules instance.
I decided to remove the underlying variable for simplicity and set a default value (empty AuthorizationRules instace) to remove the state-changing-getter that just seems to complicate things.