Better Error when a timeout occurs.
When setting a timeout on the HttpClient used by the SDK’s and a timeout occurs a TaskCanceledException is surfaced to the consumer.  Frequently when these types of timeouts occur, it is due to one of the retry handlers (RetryAfterDelegatingHandler or RetryDelegatingHandler) doing their job and retrying requests which are failing.  Ideally information about the failure which was occurring would be returned to the consumer as this is more meaningful/actionable than a TaskCanceledException.

This set of changes updates RetryAfterDelegatingHandler and RetryDelegatingHandler. To ensure that they return the last error they received when cancellation happens.  Doing this ensures that the consumer gets back a more meaningful response when the timeout is hit.