Support using MSI to fetch Key Vault access tokens (#5249)
In situations requiring cross-tenant auth, it is currently not possible
to use MSI to fetch an access token directly. In these scenarios, it
would be helpful to be able to use a client certificate grant, but with
the certificate stored in Key Vault.

Currently, this workflow is supported, but only for "development-like"
environments (i.e. through the Visual Studio or Azure CLI token flows).
This commit makes it possible for a service to use its managed identity
to retrieve a Key Vault access token for client certificate grants.