[Azure Search] Replace ContinueSearch implementation with generated code
This required some very hacky regexes to customize the generated code, but it
works. Also, it's better than having to parse URL query strings to
reconstitute a SearchParameters instance, since that would miss custom
parameters that can't be modeled in Swagger, like personalization IDs for
custom ranking.

Ideally we'll find a pattern common to other data plane APIs so we don't have
to hack this with regexes for every target language, but it's good enough for
now.