[Azure Search] Finish changing codegen strategy for extensible enums
This change is the result of correcting how we model extensible enums in the
Swagger spec. Previously, they were modeled as objects marked with
"x-ms-external": true, but this was not faithfully representing the wire
format, because on the wire these enums are strings, not objects.

Since we are not ready to fully adopt AutoRest's "extensible enums" feature,
we still have our hand-written enums, which are now structs per the previous
commit. This change adds PowerShell script to delete the newly auto-generated
enums (which are not extensible yet), but keeps the other codgen changes that
are necessary to start treating these as value types instead of reference
types. The previous commit broke the tests due to differences in nullability;
This commit fixes them.

One breaking change caused by this commit is that the SynonymMap type no
longer takes a SynonymMapFormat as a constructor parameter, and the Format
property is now a static read-only string. This is per AutoRest's policy of
"inlining" required enum parameters that have only a single possible value.