[Azure Search] Start partially generating the Field class
This change was driven by the need to update Field in the Swagger to
accurately reflect the REST API. The eventual goal is to generate Field for
SDKs in other languages. For .NET, this means replacing the fully customized
Field class with a partially-generated implementation that is mostly
backwards compatible.