[Azure Search] Make custom properties of Field nullable
We're about to add support for complex types to the SDK. This requires
modeling sub-fields of complex fields. There are two potential approaches for
this:
 1. Inheritance. Field would become a base class, with LeafField and
    ComplexField derived classes.
 2. Re-use. Field remains a concrete class, and gains a new Fields property,
    making it a recursive data type.

The way this is modeled in the REST API itself is much closer to 2 than 1.
Modeling 1 in Swagger isn't really possible without extending AutoRest's
x-ms-discriminator-value extension in unnatural ways. Rather than fight
against the nature of our API, we'll instead "go with the flow" and use
approach 2 above. This means many existing properties of Field won't be
meaningful for complex fields, and thus, must be nullable to indicate that
they do not apply.