Better errors when a timeout occurs. (#5245)
* Better Error when a timeout occurs.

When setting a timeout on the HttpClient used by the SDK’s and a timeout occurs a TaskCanceledException is surfaced to the consumer.  Frequently when these types of timeouts occur, it is due to one of the retry handlers (RetryAfterDelegatingHandler or RetryDelegatingHandler) doing their job and retrying requests which are failing.  Ideally information about the failure which was occurring would be returned to the consumer as this is more meaningful/actionable than a TaskCanceledException.

This set of changes updates RetryAfterDelegatingHandler and RetryDelegatingHandler. To ensure that they return the last error they received when cancellation happens.  Doing this ensures that the consumer gets back a more meaningful response when the timeout is hit.

* Avoid breaking binary compatability by adding new constructor overload instead of changing to base type directly.

* Reintroduce the EventCallbackCount property on the RetryDelegatingHandler to avoid breaking binary compatibility.

The property originally returned the count of handlers registered on the private Retrying EventHandler.  Since there is no longer a need for the private Retrying EventHandler, updating to return the count of handlers on the RetryPolicy since it is the closest meaningful value.

* Updating RetryAfterDelegatingHandlerTests to have a timeout in seconds now that the handler is treating the "Retry-After" in seconds.

* Bring back the CRMemoryTests

The test was previously testing that handlers are not being leaked by the RetryDelegatingHandler.

Now that the RetryDelegatingHandler is not making use of the Retrying handler, the test validates that no handlers are being registered.

* Update tests to test for null content in the response.

Avoid throwing by using when clause RetryDelegatingHandler.

* - Cleanup in RetryAfterDelegatingHandler
- Remove changes from FakeHttpHandler that were not not needed.