Update signing targets
In order to prepare for testing on .NET framework
we needed to switch some signing setting around so
that we can run the test localling without requiring
global machine setting changes to disable strong name
verification. Do do this we will use a new public signing
feature which is what the .NET team does to test these
scenarios.

- Default to using PublicSign for sdk libraries
- Enable signing for all test libraries by default
- Add xunit config file to disable shadow copying which
  doesn't work with public signing
- Disable PublicSign, thus enable DelaySigning for our package build step
  in our official builds so that we have delay-signed binaries there because
  we cannot real-sign on top of public signed binaries.
- Update template test project to actually use the template library