[Engineering System] Increase Timeout for Live Tests (#6083)
Due to the current state of the Event Hubs test suite, a live test run for
a given platform is taking roughly 90 minutes on average.  Because Windows is
running two target platforms, the test run has been averaging roughly 2 hours
and 25 minutes.

Until we're able to refactor the Event Hubs tests for test timings, isolation,
and parallelization, adjusting the timeout to allow nightly runs to complete
without timing out.