Add uap10.0 asset to the EventHubs package
Because newer versions of UWP are compatible with netstandard 2.0 we need to ensure that we provide a uap10.0 asset
in the package that doesn't depend on client certificate support. In order to do that we will copy the netstandard1.4
asset into the uap10.0 target in the package to ensure it always gets chosen when running on uwp.

Also contains some EventHubs project clean-up.