Add new public key for client libraries
For our new client libraries we want to use a full public/private
key for our libraries as we no longer use strong name identity
for security reasons and we want to make building our libraries
easier for anyone. The strong name public key is only for identity
purposes now.

The public key for AzureSDKClient is: 92742159e12e44c8