Add support for versioning packages
Adds support for versioning packages which will include
a build number based on a date that is part of the prerelease
label so we can publish daily without worrying about overwriting
an existing packages on the feed.

Also include support for releasing with a different prerelease
label or no prerelease label.

See Documentation/Versioning.md for all the details.