[Event Hubs Client] Track 2 Prototype
Refactoring sending of events to move specifying a partition to the creation of a sender,
as it was mistakenly believed that a partition could be selected at send;  this was
the batch label (aka partition key) and not the identifier.

Senders can now be created specific to a partition or non-specific, the latter of which
will allow the Event Hubs service to assume responsibility for routing of events to a
partition - the desired majority use case.