Unblock message pump while receiving message
Previous version of the code blocked further task processing with more
Tasks being created while message was being received up until the point
when message processing actually started.

With switch from await to ContinueWith execution of the rest of the
method is continued asynchronously allowing `MaxConcurrentCalls` to be
higher than number of logical processors on the target machine.