[Azure Search] Refactor FieldBuilder to use sum type for DataTypeInfo
The logic for detecting different kinds of data types will get more complex
in the future. To prepare, this change replaces the DataTypeInfo struct with
the C# equivalent of a "sum type" (a.k.a. "discriminated union"), implemented
via the IDataTypeInfo interface and its companion classes. This will allow
modeling of more mutually-exclusive cases for data types that may carry
different properties.

This change also necessitated removing one of the "continue" statements from
the inner loop. It belongs to a case that only applies to complex fields, so
logically the "continue" belongs in one of the lambdas passed to Match, but
since that's not possible, CreateComplexField just returns null for that case
instead.