[Azure Search] Make FieldBuilder fail gracefully on unknown types
We can't just assume that properties of types we don't recognize must be
complex types. Instead, FieldBuilder now uses the JSON contract resolver to
determine whether a property is of a complex type. If not, it throws an
exception with a helpful error message.

Rather than repeatedly call the contract resolver (once at the beginning of
BuildForTypeRecursive and once in GetDataTypeInfo), DataTypeInfo.Complex has
been extended to carry the JsonObjectContract for the type as a performance
optimization. The signature of BuildForTypeRecursive has been modified
accordingly so it's now the caller's responsibility to provide the
JsonObjectContract.