[Azure Search] FieldBuilder improvements (#6833)
* [Azure Search] Refactor FieldBuilder to use sum type for DataTypeInfo

The logic for detecting different kinds of data types will get more complex
in the future. To prepare, this change replaces the DataTypeInfo struct with
the C# equivalent of a "sum type" (a.k.a. "discriminated union"), implemented
via the IDataTypeInfo interface and its companion classes. This will allow
modeling of more mutually-exclusive cases for data types that may carry
different properties.

This change also necessitated removing one of the "continue" statements from
the inner loop. It belongs to a case that only applies to complex fields, so
logically the "continue" belongs in one of the lambdas passed to Match, but
since that's not possible, CreateComplexField just returns null for that case
instead.

* [Azure Search] Replacing explicit loop in FieldBuilder with function pipeline

This change cleans up the core logic of FieldBuilder, removing the last
"continue" and making it easier to reason about and extend.

* [Azure Search] Make FieldBuilder fail gracefully on unknown types

We can't just assume that properties of types we don't recognize must be
complex types. Instead, FieldBuilder now uses the JSON contract resolver to
determine whether a property is of a complex type. If not, it throws an
exception with a helpful error message.

Rather than repeatedly call the contract resolver (once at the beginning of
BuildForTypeRecursive and once in GetDataTypeInfo), DataTypeInfo.Complex has
been extended to carry the JsonObjectContract for the type as a performance
optimization. The signature of BuildForTypeRecursive has been modified
accordingly so it's now the caller's responsibility to provide the
JsonObjectContract.

* [Azure Search] Add top-level validation to FieldBuilder

Now FieldBuilder will fail with a helpful error message if a type that
clearly isn't a DTO is passed. This is determined by whether or not the type
resolves to a JsonObjectContract.

* [Azure Search] Add FieldBuilder test coverage for ICollection properties

* [Azure Search] Ignore key attribute on sub-fields

* [Azure Search] Refactor FieldBuilderTests to use fewer callbacks

Callbacks make tests harder to reason about. They were there primarily to
support calling FieldBuilder with and without a custom contract resolver.
Instead, a flag is now passed to each test to indicate whether to pass a
contract resolver to FieldBuilder.

Also, the tests now more explicitly call FieldBuilder and then assert on
the result using a new helper class, FieldMap. This makes the data flow in
the tests more obvious and more closely follows the arrange-act-assert
pattern.

* [Azure Search] Add missing docs to FieldBuilder-related attributes

* [Azure Search] Add new FieldBuilderIgnoreAttribute

Sometimes model classes have properties of types (like enums) that make sense
for application code, but don't automatically map to an Azure Search data
type. Using [JsonIgnore] is not an option in scenarios when such properties
actually are a part of the index. This commit introduces [FieldBuilderIgnore]
which FieldBuilder treats just like [JsonIgnore]. That way, you can opt out
of FieldBuilder without opting out of JSON serialization/deserialization.

* [Azure Search] Factor out data plane version into props file

* Update session record for live FieldBuilder test

* [Azure Search] Replace CRLF with Environment.NewLine

* [Azure Search] Rename a parameter per PR feedback